Add files via upload
Experiment:
Design a circuit for interfacing the LDR Sensor and print the value of light intensity at various
instances.
Design a circuit for interfacing PIR Sensor and discuss the use of status variable in the sketch.

Submission on LMS: Following point you must submit in the assignment.
1. Problem description required Components and circuit diagram.
2. Explain the concepts, experimental, and programming skills that you have gained during this
experiment. Discuss the results and comment on the reasons for the errors.
3. What is the maximum and minimum Analog Digital value of LDR Sensor? Write five real time
applications (not from internet) of both the sensors.